refactor: update path for team data file
