+++ src/config/content/mycomp_data/team.js
